+++ SDK/AppCenter/Microsoft.AppCenter.Windows.Shared/Ingestion/Http/NetworkIngestionException.cs
@@ -2,13 +2,14 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Microsoft.AppCenter.Ingestion.Http
 {
     public class NetworkIngestionException : IngestionException
     {
         public override bool IsRecoverable => true;
+
+        public NetworkIngestionException() { }
+        public NetworkIngestionException(Exception innerException) : base(innerException) { }
     }
 }